[cms] Download from library uses wrong filename [cms] Display group should delete permissions

+++ server/lib/data/displaygroup.data.class.php
@@ -151,7 +151,8 @@ class DisplayGroup extends Data
      */
     public function Delete($displayGroupID)
     {
-        Debug::LogEntry('audit', 'IN', 'DisplayGroup', 'Delete');
+        if ($displayGroupID == NULL || $displayGroupID == 0)
+            return $this->SetError(__('Missing displayGroupId'));
 
         try {
             $dbh = PDOConnect::init();
@@ -161,6 +162,13 @@ class DisplayGroup extends Data
     
             if (!$schedule->DeleteScheduleForDisplayGroup($displayGroupID))
                 throw new Exception('Unable to DeleteScheduleForDisplayGroup');
+
+            // Remove all permissions
+            Kit::ClassLoader('displaygroupsecurity');
+            $security = new DisplayGroupSecurity($this->db);
+
+            if (!$security->UnlinkAll($displayGroupID))
+                throw new Exception('Unable to Unlink all Display Group Permissions');                
 
             // Delete the Display Group
             $sth = $dbh->prepare('DELETE FROM displaygroup WHERE DisplayGroupID = :displaygroupid');
